requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pymongo>=4.6
python-dotenv>=1.0
python-multipart>=0.0.9
pydantic>=2.6
passlib[bcrypt]
